pyproject.toml
[project]
name = "vidflow"
description = "Unified video capture and transcription CLI"
authors = [
    { name = "Joe Monaco", email = "joe@selfmotion.net" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["video", "transcription", "youtube", "claude", "obsidian", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dynamic = ["version"]

dependencies = [
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "Pillow>=10.0.0",
    # capture (formerly ytcapture)
    "yt-dlp>=2024.0.0",
    "youtube-transcript-api>=0.6.0",
    "python-dateutil>=2.8.0",
    "imagehash>=4.3.0",
    # transcribe (formerly vidscribe)
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
ai = ["anthropic>=0.40.0"]
citations = ["exa-py>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
vidflow = "vidflow.cli:main"
ytcapture = "vidflow.capture.cli:ytcapture_entry"
vidcapture = "vidflow.capture.cli:vidcapture_entry"
vidscribe = "vidflow.transcribe.cli:vidscribe_entry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/vidflow/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/vidflow"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
